requirements.txt
Flask==0.8
unicodecsv
pillow
SQLAlchemy==0.7.8
Flask-SQLAlchemy==1.0
psycopg2==2.4.5
